Merged PR 1091: Fix for routing block race disposal issue
Routing block was disposing of routes and the DI scopes for those routes whilst downstream blocks in the route were still executing and attmepting to use the scope.

This PR fixes this, and adds more regression test coverage for this issue.

In addition to this some simple benchmakrs were added.

Related work items: #9719